Make recentering point at window edge optional
Sublimity mimics the default Emacs behavior when the point reaches the edge of the window. However, it would be nice to have smooth-scrolling behave exactly like Sublime such that the window only scrolls one line at a time when moving the point outside the window.

This leaves the default behavior, but lets you disable the `(recenter)` call so that when someone changes Emacs' scrolling behavior (i.e. `(setq scroll-conservatively 101`) it behaves like you'd expect.

Is this an okay solution?